+++ vast_pipeline/image/utils.py
@@ -253,13 +253,22 @@ def calc_condon_flux_errors(
         return 0., 0., 0., 0., 0., 0., 0.
 
 
-def open_fits(fits_path: Union[str, Path], memmap: Optional[bool] = True):
+def open_fits(
+    fits_path: Union[str, Path],
+    memmap: Optional[bool] = True,
+    comp_nan_fill: Optional[bool]= True,
+    comp_nan_fill_cut = -1e4,
+) -> fits.HDUList:
     """
     This function opens both compressed and uncompressed fits files.
 
     Args:
         fits_path: Path to the fits file
         memmap: Open the fits file with mmap.
+        comp_nan_fill: Fill formerly-NaN values with NaNs in compressed images.
+            Defaults to True.
+        comp_nan_fill_cut: The cutoff value for replacing negative numbers
+            with NaNs. Only relevant if `comp_nan_fill=True`. Defaults to -1e4.
 
     Returns:
         HDUList loaded from the fits file
@@ -273,7 +282,10 @@ def open_fits(fits_path: Union[str, Path], memmap: Optional[bool] = True):
     # This is a messy way to check, but I can't think of a better one
     if len(hdul) == 1:
         return hdul
-    elif type(hdul[1]) == fits.hdu.compressed.CompImageHDU:
+    elif isinstance(hdul[1], fits.hdu.compressed.CompImageHDU):
+        if comp_nan_fill:
+            data = hdul[1].data
+            data[data<comp_nan_fill_cut] = np.nan
         return fits.HDUList(hdul[1:])
     else:
         return hdul